requirements.txt
googletrans==4.0.0-rc1
pillow~=9.4.0
transformers~=4.29.2
mitmproxy~=10.0.0
pyotp~=2.9.0
beautifulsoup4~=4.12.2
numpy~=1.24.3